LayoutTests/http/tests/media/media-source/video-media-source-state-changes.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/media-resources/video-test.js"></script>
        <script src="/media/resources/media-source/webm/segment-info.js"></script>
        <script src="media-source.js"></script>
        <script>
            var segmentHelper = new MediaSourceTest.SegmentHelper(WebMSegmentInfo.testWebM);
            var seekTime = 1;

            function appendMediaSegmentsForSeek(segmentSeekTime)
            {
                var index = segmentHelper.getMediaSegmentIndexForTimestamp(segmentSeekTime);
                var endIndex = index + 2;
                var totalMediaSegments = segmentHelper.mediaSegments.length;

                for (; (index < endIndex) && (index < totalMediaSegments); index++) {
                    segmentHelper.appendMediaSegment(index);
                }

                if (index >= totalMediaSegments) {
                    mediaSource.endOfStream();
                    return;
                }
            }

            function onSourceEvent(event)
            {
                consoleWrite('EVENT(' + event.type + ') : ' + event.target.readyState);
            }

            function onFirstSourceOpen(event)
            {
                consoleWrite("onFirstSourceOpen");

                // This is the first time the source is opened. We just want to append
                // the initialization segment and all media segments until the end of stream.
                // This is testing the
                // "closed" -> "open" -> "ended" transition path.

                var mediaSource = event.target;
                mediaSource.removeEventListener('webkitsourceopen', onFirstSourceOpen);
                segmentHelper.addSourceBuffer();

                MediaSourceTest.expectSourceState(mediaSource, "open");

                segmentHelper.appendInitSegment();

                mediaSource.addEventListener('webkitsourceended', onFirstSourceEnded);

                // Start appending media segments.
                segmentHelper.appendUntilEndOfStream(0);
            }

            function onFirstSourceEnded(event)
            {
                consoleWrite("onFirstSourceEnded");

                var mediaSource = event.target;
                mediaSource.removeEventListener('webkitsourceended', onFirstSourceEnded);

                MediaSourceTest.expectSourceState(mediaSource, "ended");

                video.addEventListener('playing', triggerFirstSeek);
            }

            function triggerFirstSeek(event)
            {
                consoleWrite("triggerFirstSeek");

                var videoTag = event.target;
                videoTag.removeEventListener('playing', triggerFirstSeek);

                // Now that we have appended all of the media segments and ended the stream,
                // lets do a seek to make sure we stay in the "ended" state.
                videoTag.addEventListener('seeking', onFirstSeeking);
                videoTag.addEventListener('seeked', onFirstSeeked);
                videoTag.currentTime = seekTime;
            }

            function onFirstSeeking(event)
            {
                consoleWrite("onFirstSeeking");

                var videoTag = event.target;
                videoTag.removeEventListener('seeking', onFirstSeeking);

                MediaSourceTest.expectSourceState(mediaSource, "ended");

                // Start appending media segments.
                appendMediaSegmentsForSeek(videoTag.currentTime);

                // Verify that the source state transitioned to open after the append.
                MediaSourceTest.expectSourceState(mediaSource, "open");
            }

            function onFirstSeeked(event)
            {
                consoleWrite("onFirstSeeked");

                var videoTag = event.target;
                videoTag.removeEventListener('seeked', onFirstSeeked);

                MediaSourceTest.expectSourceState(mediaSource, "open");

                // Now it is time to test seeking while still in "open".
                videoTag.addEventListener('seeking', onSecondSeeking);
                videoTag.addEventListener('seeked', onSecondSeeked);
                videoTag.currentTime = seekTime;
            }

            function onSecondSeeking(event)
            {
                consoleWrite("onSecondSeeking");

                var videoTag = event.target;
                videoTag.removeEventListener('seeking', onSecondSeeking);

                MediaSourceTest.expectSourceState(mediaSource, "open");

                // Start appending media segments.
                appendMediaSegmentsForSeek(videoTag.currentTime);
            }

            function onSecondSeeked(event)
            {
                consoleWrite("onSecondSeeked");

                var videoTag = event.target;
                videoTag.removeEventListener('seeked', onSecondSeeked);

                MediaSourceTest.expectSourceState(mediaSource, "open");

                // Test "open" -> "closed" -> "open" transition path.
                mediaSource.addEventListener('webkitsourceclose', onFirstSourceClose);
                MediaSourceTest.setSrcToMediaSourceTestURL(videoTag);
            }

            function onFirstSourceClose(event)
            {
                consoleWrite("onFirstSourceClose");

                var mediaSource = event.target;
                mediaSource.removeEventListener('webkitsourceclose', onFirstSourceClose);

                MediaSourceTest.expectSourceState(mediaSource, "closed");

                mediaSource.addEventListener('webkitsourceopen', onSecondSourceOpen);
            }

            function onSecondSourceOpen(event)
            {
                consoleWrite("onSecondSourceOpen");

                var mediaSource = event.target;
                mediaSource.removeEventListener('webkitsourceopen', onSecondSourceOpen);
                segmentHelper.addSourceBuffer();

                MediaSourceTest.expectSourceState(mediaSource, "open");

                segmentHelper.appendInitSegment();

                mediaSource.addEventListener('webkitsourceended', onSecondSourceEnded);

                // Start appending media segments.
                segmentHelper.appendUntilEndOfStream(0);
            }

            function onSecondSourceEnded(event)
            {
                consoleWrite("onSecondSourceEnded");

                var mediaSource = event.target;
                mediaSource.removeEventListener('webkitsourceended', onSecondSourceEnded);

                MediaSourceTest.expectSourceState(mediaSource, "ended");
                //MediaSourceTest.expectReadyState(videoTag, HTMLMediaElement.HAVE_NOTHING);

                video.addEventListener('playing', triggerSecondSourceClose);
            }

            function triggerSecondSourceClose(event)
            {
                consoleWrite("triggerSecondSourceClose");

                var videoTag = event.target;
                videoTag.removeEventListener('playing', triggerSecondSourceClose);

                // Test "ended" -> "closed" -> "open" transition path.
                mediaSource.addEventListener('webkitsourceclose', onSecondSourceClose);
                MediaSourceTest.setSrcToMediaSourceTestURL(videoTag);
            }

            function onSecondSourceClose(event)
            {
                consoleWrite("onSecondSourceClose");

                var mediaSource = event.target;
                mediaSource.removeEventListener('webkitsourceclose', onSecondSourceClose);
                MediaSourceTest.expectSourceState(mediaSource, "closed");

                mediaSource.addEventListener('webkitsourceopen', onThirdSourceOpen);
            }

            function onThirdSourceOpen(event)
            {
                consoleWrite("onThirdSourceOpen");

                var mediaSource = event.target;
                mediaSource.removeEventListener('webkitsourceopen', onThirdSourceOpen);
                segmentHelper.addSourceBuffer();

                MediaSourceTest.expectSourceState(mediaSource, "open");

                endTest();
            }

            function onLoad()
            {
                findMediaElement();

                mediaSource = new WebKitMediaSource();

                waitForEventAndFail('error');
                waitForEvent('playing');
                waitForEvent('seeking');
                waitForEvent('seeked');
                waitForEvent('ended');
                waitForEvent('emptied');

                mediaSource.addEventListener('webkitsourceopen', onSourceEvent);
                mediaSource.addEventListener('webkitsourceended', onSourceEvent);
                mediaSource.addEventListener('webkitsourceclose', onSourceEvent);

                mediaSource.addEventListener('webkitsourceopen', onFirstSourceOpen);

                MediaSourceTest.logSourceState(mediaSource);
                segmentHelper.init(video, function(success)
                {
                    if (!success) {
                        failTest("Failed to load segment data");
                        return;
                    }

                    MediaSourceTest.setSrcToMediaSourceTestURL(video);
                });
            }
        </script>
    </head>
    <body onload="onLoad()">
        <video autoplay> </video>
        <p>Tests MediaSource API state transitions.</p>
    </body>
</html>
